LPS-98855 Reorganize folders
