+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/components/App.es.js
@@ -15,8 +15,8 @@
 import React from 'react';
 import {Route, HashRouter as Router, Switch} from 'react-router-dom';
 import CustomObject from './custom-object/CustomObject.es';
+import {CUSTOM_OBJECTS} from './search-container/constants.es';
 import SearchContainer from './search-container/SearchContainer.es';
-import {CUSTOM_OBJECTS} from '../utils/constants.es';
 
 export default function App() {
 	return (
