+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/components/custom-object/CustomObject.es.js
@@ -15,8 +15,8 @@
 import React, {Fragment} from 'react';
 import {Route, Switch} from 'react-router-dom';
 import NavigationBar from './NavigationBar.es';
+import {APPS, FORM_VIEWS, TABLE_VIEWS} from '../search-container/constants.es';
 import SearchContainer from '../search-container/SearchContainer.es';
-import {APPS, FORM_VIEWS, TABLE_VIEWS} from '../../utils/constants.es';
 
 export default function CustomObject({
 	match: {
